move DumpImages from restorable to graphicscommand
