+++ internal/graphicscommand/image_js.go
@@ -15,29 +15,16 @@
 package graphicscommand
 
 import (
+	"archive/zip"
 	"bytes"
 	"image"
-	"path/filepath"
 	"syscall/js"
 
 	"github.com/hajimehoshi/ebiten/v2/internal/graphicsdriver"
 )
 
-func (i *Image) Dump(graphicsDriver graphicsdriver.Graphics, path string, blackbg bool, rect image.Rectangle) error {
-	// Screen image cannot be dumped.
-	if i.screen {
-		return nil
-	}
-
-	path = i.DumpName(path)
-	path = filepath.Base(path)
-
+func jsDownload(buf *bytes.Buffer, mime string, path string) {
 	global := js.Global()
-
-	buf := &bytes.Buffer{}
-	if err := i.DumpTo(buf, graphicsDriver, blackbg, rect); err != nil {
-		return err
-	}
 
 	jsData := global.Get("Uint8Array").New(buf.Len())
 	js.CopyBytesToJS(jsData, buf.Bytes())
@@ -45,11 +32,50 @@ func (i *Image) Dump(graphicsDriver graphicsdriver.Graphics, path string, blackb
 	a := global.Get("document").Call("createElement", "a")
 	blob := global.Get("Blob").New(
 		[]interface{}{jsData},
-		map[string]interface{}{"type": "image/png"},
+		map[string]interface{}{"type": mime},
 	)
 	a.Set("href", global.Get("URL").Call("createObjectURL", blob))
 	a.Set("download", path)
 	a.Call("click")
+}
+
+func (i *Image) Dump(graphicsDriver graphicsdriver.Graphics, path string, blackbg bool, rect image.Rectangle) error {
+	// Screen image cannot be dumped.
+	if i.screen {
+		return nil
+	}
+
+	buf := &bytes.Buffer{}
+	if err := i.dumpTo(buf, graphicsDriver, blackbg, rect); err != nil {
+		return err
+	}
+
+	jsDownload(buf, "image/png", i.dumpName(path))
+
+	return nil
+}
+
+// DumpImages dumps all the specified images to the specified directory.
+//
+// This is for testing usage.
+func DumpImages(images []*Image, graphicsDriver graphicsdriver.Graphics, dir string) error {
+	buf := &bytes.Buffer{}
+	zw := zip.NewWriter(buf)
+
+	for _, img := range images {
+		f, err := zw.Create(img.dumpName("*.png"))
+		if err != nil {
+			return err
+		}
+
+		if err := img.dumpTo(f, graphicsDriver, false, image.Rect(0, 0, img.width, img.height)); err != nil {
+			return err
+		}
+	}
+
+	zw.Close()
+
+	jsDownload(buf, "archive/zip", dir+".zip")
 
 	return nil
 }
