--- internal/restorable/images_notjs.go
@@ -1,42 +1,0 @@
-// Copyright 2022 The Ebitengine Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build !js
-// +build !js
-
-package restorable
-
-import (
-	"image"
-	"os"
-	"path/filepath"
-
-	"github.com/hajimehoshi/ebiten/v2/internal/graphicsdriver"
-)
-
-// DumpImages dumps all the current images to the specified directory.
-//
-// This is for testing usage.
-func DumpImages(graphicsDriver graphicsdriver.Graphics, dir string) error {
-	if err := os.Mkdir(dir, 0755); err != nil {
-		return err
-	}
-
-	for img := range theImages.images {
-		if err := img.Dump(graphicsDriver, filepath.Join(dir, "*.png"), false, image.Rect(0, 0, img.width, img.height)); err != nil {
-			return err
-		}
-	}
-	return nil
-}
